create shared config
