+++ client/configuration/webpack/webpack.config.dev.js
@@ -2,31 +2,23 @@ const path = require("path");
 const webpack = require("webpack");
 const HtmlWebpackPlugin = require("html-webpack-plugin");
 const FaviconsWebpackPlugin = require("favicons-webpack-plugin");
-const ObsoleteWebpackPlugin = require("obsolete-webpack-plugin");
 const ScriptExtHtmlWebpackPlugin = require("script-ext-html-webpack-plugin");
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 
-const src = path.resolve("src");
+const { merge } = require("webpack-merge");
+
+const sharedConfig = require("./webpack.config.shared");
+const babelOptions = require("../babel/babel.dev");
+
 const fonts = path.resolve("src/fonts");
 const nodeModules = path.resolve("node_modules");
 
-const babelOptions = require("../babel/babel.dev");
-
-module.exports = {
+const devConfig = {
   mode: "development",
   devtool: "eval",
-  entry: [
-    "core-js",
-    "regenerator-runtime/runtime",
-    "fastestsmallesttextencoderdecoder",
-    "whatwg-fetch",
-    "abort-controller/polyfill",
-    "./src/index",
-  ],
   output: {
-    path: path.resolve("build"),
     pathinfo: true,
     filename: "static/js/bundle.js",
-    publicPath: "/",
   },
   module: {
     rules: [
@@ -35,38 +27,6 @@ module.exports = {
         loader: "babel-loader",
         options: babelOptions,
       },
-      {
-        test: /\.css$/,
-        include: src,
-        exclude: [path.resolve(src, "index.css")],
-        loader: [
-          {
-            loader: "style-loader",
-          },
-          {
-            loader: "css-loader",
-            options: {
-              modules: {
-                localIdentName: "[name]__[local]___[hash:base64:5]",
-              },
-            },
-          },
-        ],
-      },
-      {
-        test: /index\.css$/,
-        include: [path.resolve(src, "index.css")],
-        loader: [
-          {
-            loader: "style-loader",
-          },
-          {
-            loader: "css-loader",
-          },
-        ],
-      },
-
-      { test: /\.json$/, include: [src, nodeModules], loader: "json-loader" },
       {
         test: /\.(jpg|png|gif|eot|svg|ttf|woff|woff2|otf)$/i,
         loader: "file-loader",
@@ -79,44 +39,6 @@ module.exports = {
     new HtmlWebpackPlugin({
       inject: true,
       template: path.resolve("index.html"),
-    }),
-    new ObsoleteWebpackPlugin({
-      name: "obsolete",
-      template:
-        "<script>" +
-        'var root = document.getElementById("root");' +
-        "root.remove();" +
-        'var portals = document.getElementsByClassName("bp3-portal");' +
-        "for(var i = 0; i < portals.length; i += 1) {" +
-        "portals[i].remove();" +
-        "}" +
-        "</script>" +
-        '<div style="display: flex;flex-direction: column;width: 100vw;height: 100vh;text-align: center;justify-content: center;align-items: center;background: #8080801a;font-family: &quot;Roboto Condensed, sans serif&quot;;">' +
-        '<img src="https://raw.githubusercontent.com/chanzuckerberg/cellxgene/main/docs/cellxgene-logo.png" style="width: 320px;"/>' +
-        '<div style="margin-top: 16px;background: white;width: 40vw;border-radius: 4px;padding: 24px 64px;-webkit-box-shadow: 0px 0px 3px 2px rgba(0,0,0,0.38);-moz-box-shadow: 0px 0px 3px 2px rgba(0,0,0,0.38);box-shadow: 0px 0px 3px 2px rgba(0,0,0,0.38);max-width: 550px;">' +
-        '<div style="margin-bottom: 0;font-weight: bolder;font-size: 1.2em;"> Unsupported Browser </div>' +
-        '<div style="margin-top: 0;">cellxgene is currently supported on the following browsers</div>' +
-        '<div style="display: flex;justify-content: space-around;margin-top: 16px">' +
-        '<a href="https://www.google.com/chrome/?hl=en%22" aria-label="Download Google Chrome">' +
-        '<img src="https://assets.beta.meta.org/images/browsers/chrome.png" style="width: 80px;height: 80px;"/>' +
-        "<div>Chrome &gt; 60</div>" +
-        "</a>" +
-        '<a href="https://www.apple.com/safari/" aria-label="Download Safari">' +
-        '<img src="https://assets.beta.meta.org/images/browsers/safari.png" style="width: 80px;height: 80px;"/>' +
-        "<div>Safari ≥ 10.1</div>" +
-        "</a>" +
-        '<a href="https://www.mozilla.com/firefox/" aria-label="Download Firefox">' +
-        '<img src="https://assets.beta.meta.org/images/browsers/firefox.png" style="width: 80px;height: 80px;"/>' +
-        "<div>Firefox ≥ 60</div>" +
-        "</a>" +
-        '<a href="//www.microsoft.com/edge" aria-label="Download Edge">' +
-        '<img src="https://assets.beta.meta.org/images/browsers/edge.png" style="width: 80px;height: 80px;"/>' +
-        "<div>Edge ≥ 15</div>" +
-        "</a>" +
-        "</div>" +
-        "</div>" +
-        "</div>",
-      promptOnNonTargetBrowser: false,
     }),
     new FaviconsWebpackPlugin({
       logo: "./favicon.png",
@@ -133,6 +55,9 @@ module.exports = {
         },
       },
     }),
+    new MiniCssExtractPlugin({
+      filename: "static/[name].css",
+    }),
     new webpack.NoEmitOnErrorsPlugin(),
     new webpack.DefinePlugin({
       __REACT_DEVTOOLS_GLOBAL_HOOK__: "({ isDisabled: true })",
@@ -147,3 +72,5 @@ module.exports = {
     }),
   ],
 };
+
+module.exports = merge(sharedConfig, devConfig);
